datagen/requirements.txt
sseclient-py==1.8.0
kafka-python==2.0.2
